Update README.md
This would not run on my titanium. And I couldn’t find any additional error. The code only had 5 colors of teas. After moving around some semicolons and parentheses, the code run a little better but still wouldn’t show in the emulator.  I’m not sure how to fix this issue. There are two window navigation points in the code but it does not seem clear if they are linked correctly. I would look over where you opened and closed your windows to ensure that they are linked. a.	The phone gadget used was the camera app. There were no errors in that portion of the code and when ran alone worked perfectly. Good work, just needs to be polished a bit.